research-agent/hello.py
import os
from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()

client = OpenAI(
    api_key=os.getenv("DEEPSEEK_API_KEY"),
    base_url="https://api.deepseek.com",
)

response = client.chat.completions.create(
    model="deepseek-chat",
    max_tokens=200,
    messages=[
        {"role": "user", "content": "In one sentence, what is an AI agent?"}
    ],
)

print(response.choices[0].message.content)