research-agent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.97.0",
    "httpx>=0.28.1",
    "openai>=2.33.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
